fixed vertex circulator, changed laplacian computations

+++ color.h
@@ -1,8 +1,10 @@
 #ifndef COLOR_H
 #define COLOR_H
 
+#include<cmath>
 #include <array>
+#include <iostream>
 
-std::array<int, 3> hsv2rgb(double h, double s, double v);
+std::array<double, 3> hsv2rgb(int h, double s, double v);
 
 #endif // COLOR_H
